refactor(dashboard): rename insights-sidebar to insights-dock
The component stopped being a sidebar — it's a docked composer that
expands in place into the chat panel. Rename the module, update all
importers, and fix comments that still described the old sidebar
behavior. No functional changes.
